+++ src/spyglass/lfp/lfp_imported.py
@@ -14,7 +14,7 @@ class ImportedLFP(SpyglassMixin, dj.Imported):
     definition = """
     -> Session                      # the session to which this LFP belongs
     -> LFPElectrodeGroup            # the group of electrodes to be filtered
-    -> IntervalList # the original set of times to be filtered
+    -> IntervalList                 # the original set of times to be filtered
     lfp_object_id: varchar(40)      # object ID for loading from the NWB file
     ---
     lfp_sampling_rate: float        # the sampling rate, in samples/sec
@@ -22,6 +22,7 @@ class ImportedLFP(SpyglassMixin, dj.Imported):
     """
 
     def make(self, key):
+        """Placeholder for importing LFP."""
         raise NotImplementedError(
             "For `insert`, use `allow_direct_insert=True`"
         )