+++ src/spyglass/utils/dj_graph.py
@@ -13,7 +13,7 @@ from typing import Any, Dict, Iterable, List, Set, Tuple, Union
 
 from datajoint import FreeTable, Table
 from datajoint.condition import make_condition
-from datajoint.dependencies import unite_master_parts
+from datajoint.hash import key_hash
 from datajoint.user_tables import TableMeta
 from datajoint.utils import get_master, to_camel_case
 from networkx import (
@@ -33,6 +33,11 @@ from spyglass.utils.dj_helper_fn import (
     fuzzy_get,
     unique_dicts,
 )
+
+try:  # Datajoint 0.14.2+ uses topo_sort instead of unite_master_parts
+    from datajoint.dependencies import topo_sort as dj_topo_sort
+except ImportError:
+    from datajoint.dependencies import unite_master_parts as dj_topo_sort
 
 
 class Direction(Enum):
@@ -473,7 +478,7 @@ class AbstractGraph(ABC):
             if not self._is_out(node, warn=False)
         ]
         graph = self.graph.subgraph(nodes) if subgraph else self.graph
-        ordered = unite_master_parts(list(topological_sort(graph)))
+        ordered = dj_topo_sort(list(topological_sort(graph)))
         if reverse:
             ordered.reverse()
         return [n for n in ordered if n in nodes]
@@ -601,7 +606,8 @@ class RestrGraph(AbstractGraph):
         """Return hash of all visited nodes."""
         initial = hash_md5(b"")
         for table in self.all_ft:
-            initial.update(table.fetch())
+            for row in table.fetch(as_dict=True):
+                initial.update(key_hash(row).encode("utf-8"))
         return initial.hexdigest()
 
     # ------------------------------- Add Nodes -------------------------------
@@ -939,6 +945,7 @@ class TableChain(RestrGraph):
 
     @property
     def path_str(self) -> str:
+        """Return string representation of path: parent -> {links} -> child."""
         if not self.path:
             return "No link"
         return self._link_symbol.join([self._camel(t) for t in self.path])
@@ -981,6 +988,7 @@ class TableChain(RestrGraph):
     # ---------------------------- Graph Traversal ----------------------------
 
     def cascade_search(self) -> None:
+        """Cascade restriction through graph to search for applicable table."""
         if self.cascaded:
             return
         restriction, restr_attrs = self._get_find_restr(self.leaf)
@@ -1033,6 +1041,7 @@ class TableChain(RestrGraph):
         limit: int = 100,
         **kwargs,
     ):
+        """Search parents/children for a match of the provided restriction."""
         if (
             self.found_restr
             or not table
@@ -1135,6 +1144,7 @@ class TableChain(RestrGraph):
     def cascade(
         self, restriction: str = None, direction: Direction = None, **kwargs
     ):
+        """Cascade restriction up or down the chain."""
         if not self.has_link:
             return