+++ tests/conftest.py
@@ -394,6 +394,7 @@ def frequent_imports():
     from spyglass.lfp.analysis.v1 import LFPBandSelection
     from spyglass.mua.v1.mua import MuaEventsV1
     from spyglass.ripple.v1.ripple import RippleTimesV1
+    from spyglass.spikesorting.analysis.v1.unit_annotation import UnitAnnotation
     from spyglass.spikesorting.v0.figurl_views import SpikeSortingRecordingView
 
     return (
@@ -403,6 +404,7 @@ def frequent_imports():
         RippleTimesV1,
         SortedSpikesIndicatorSelection,
         SpikeSortingRecordingView,
+        UnitAnnotation,
         UnitMarksIndicatorSelection,
     )
 
@@ -465,8 +467,6 @@ def trodes_params(trodes_params_table, teardown):
         [v for k, v in paramsets.items()], skip_duplicates=True
     )
     yield paramsets
-    if teardown:
-        trodes_params_table.delete(safemode=False)
 
 
 @pytest.fixture(scope="session")
@@ -488,8 +488,6 @@ def trodes_sel_keys(
     ]
     trodes_sel_table.insert(keys, skip_duplicates=True)
     yield keys
-    if teardown:
-        trodes_sel_table.delete(safemode=False)
 
 
 @pytest.fixture(scope="session")
@@ -497,8 +495,6 @@ def trodes_pos_v1(teardown, sgp, trodes_sel_keys):
     v1 = sgp.v1.TrodesPosV1()
     v1.populate(trodes_sel_keys)
     yield v1
-    if teardown:
-        v1.delete(safemode=False)
 
 
 @pytest.fixture(scope="session")
@@ -609,8 +605,6 @@ def track_graph(teardown, sgpl, track_graph_key):
     )
 
     yield sgpl.TrackGraph & {"track_graph_name": "6 arm"}
-    if teardown:
-        sgpl.TrackGraph().delete(safemode=False)
 
 
 @pytest.fixture(scope="session")
@@ -645,8 +639,6 @@ def lin_sel(teardown, sgpl, lin_sel_key):
     sel_table = sgpl.LinearizationSelection()
     sel_table.insert1(lin_sel_key, skip_duplicates=True)
     yield sel_table
-    if teardown:
-        sel_table.delete(safemode=False)
 
 
 @pytest.fixture(scope="session")
@@ -654,8 +646,6 @@ def lin_v1(teardown, sgpl, lin_sel):
     v1 = sgpl.LinearizedPositionV1()
     v1.populate()
     yield v1
-    if teardown:
-        v1.delete(safemode=False)
 
 
 @pytest.fixture(scope="session")
@@ -816,6 +806,13 @@ def dlc_project_name():
 
 
 @pytest.fixture(scope="session")
+def team_name(common):
+    team_name = "sc_eb"
+    common.LabTeam.insert1({"team_name": team_name}, skip_duplicates=True)
+    yield team_name
+
+
+@pytest.fixture(scope="session")
 def insert_project(
     verbose_context,
     teardown,
@@ -823,6 +820,7 @@ def insert_project(
     dlc_project_name,
     dlc_project_tbl,
     common,
+    team_name,
     bodyparts,
     mini_copy_name,
 ):
@@ -845,8 +843,6 @@ def insert_project(
         RippleTimesV1,
     )
 
-    team_name = "sc_eb"
-    common.LabTeam.insert1({"team_name": team_name}, skip_duplicates=True)
     video_list = common.VideoFile().fetch(
         "nwb_file_name", "epoch", as_dict=True
     )[:2]
@@ -882,7 +878,6 @@ def insert_project(
     yield project_key, cfg, config_path
 
     if teardown:
-        (dlc_project_tbl & project_key).delete(safemode=False)
         shutil_rmtree(str(Path(config_path).parent))
 
 
@@ -989,13 +984,16 @@ def populate_training(
     if len(train_tbl & model_train_key) == 0:
         _ = add_training_files
         DOWNLOADS.move_dlc_items(labeled_vid_dir)
-        sgp.v1.DLCModelTraining.populate(model_train_key)
+    sgp.v1.DLCModelTraining().populate(model_train_key)
     yield model_train_key
 
 
 @pytest.fixture(scope="session")
 def model_source_key(sgp, model_train_key, populate_training):
-    yield (sgp.v1.DLCModelSource & model_train_key).fetch1("KEY")
+
+    _ = populate_training
+
+    yield (sgp.v1.DLCModelSource & model_train_key).fetch("KEY")[0]
 
 
 @pytest.fixture(scope="session")